pyproject.toml
[tool.poetry]
name = "django-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
pytest-django = "^4.5.2"
pytest = "^7.4.2"

[tool.poetry.group.dev.dependencies]
celery = "^5.3.4"
djangorestframework = "^3.14.0"
redis = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
